feat(wave7): audio and screen recording, and the media they live in
Microphone through cpal, system sound through WASAPI loopback, and the screen
through Windows Graphics Capture — a chosen monitor or a chosen window, encoded
to H.264 in the same call. The library lists recordings with duration, size,
playback, rename and delete.

Rust picks the file path; JS never passes one. recording_start creates the file
inside the media directory itself and returns it, so the IPC surface has no
user-supplied path to guard.

Playback could not open a single file, and the reason was Tauri's, not ours: the
asset-protocol scope variable $EXE is unsupported on Windows, and a portable
build keeps media beside the executable, so no static scope could ever name it.
Media is served by a tempo-media scheme that resolves the directory at runtime
and honours Range, which is what makes video seek.

Three defects came out of the live run and none of them from a test:

Every recording failed to reach the library. duration_sec was declared integer
while a recording reports 4.21 s, so the insert was refused. The column is Real
now, for sessions too: a duration is fractional by nature.

mic: None meant "no microphone" rather than "the default one". The UI sends
undefined until the device list has loaded, so a recording could silently have
no sound at all.

The screen-capture confirmation was labelled "Screen capture unavailable",
which reads as a disabled feature rather than a question.

Measured on this machine: microphone 4.28 s at peak 0.30; loopback 0.20 while a
sound played against 0.005 in silence; screen 1920x1080 at 8.06% of one core
inside a 15% budget, in a debug build.

+++ src-tauri/src/storage/migrations.rs
@@ -61,7 +61,9 @@ CREATE TABLE IF NOT EXISTS drawings (id TEXT PRIMARY KEY, title TEXT, scene_json
   preview_path TEXT, updated_at TEXT NOT NULL, deleted_at TEXT);
 CREATE TABLE IF NOT EXISTS recordings (id TEXT PRIMARY KEY, title TEXT,
   kind TEXT NOT NULL,                       -- audio | screen
-  file_path TEXT NOT NULL, duration_sec INTEGER, transcript TEXT, transcript_status TEXT,
+  -- Note: duration_sec is declared REAL (floating point seconds). Existing installs keep an INTEGER-affinity
+  -- column where SQLite stores fractional values unchanged; no table migration is needed.
+  file_path TEXT NOT NULL, duration_sec REAL, transcript TEXT, transcript_status TEXT,
   updated_at TEXT NOT NULL, deleted_at TEXT);
 CREATE TABLE IF NOT EXISTS events (id TEXT PRIMARY KEY, source TEXT NOT NULL,   -- local | google
   google_id TEXT, calendar_id TEXT, title TEXT, start_at TEXT, end_at TEXT,
@@ -69,7 +71,7 @@ CREATE TABLE IF NOT EXISTS events (id TEXT PRIMARY KEY, source TEXT NOT NULL,   
   updated_at TEXT NOT NULL, deleted_at TEXT);
 CREATE TABLE IF NOT EXISTS calendars_meta (calendar_id TEXT PRIMARY KEY, sync_token TEXT, last_sync_at TEXT);
 CREATE TABLE IF NOT EXISTS sessions (id TEXT PRIMARY KEY, kind TEXT NOT NULL,    -- pomodoro | stopwatch
-  started_at TEXT, ended_at TEXT, duration_sec INTEGER, completed INTEGER, task_id TEXT,
+  started_at TEXT, ended_at TEXT, duration_sec REAL, completed INTEGER, task_id TEXT,
   updated_at TEXT NOT NULL, deleted_at TEXT);
 CREATE TABLE IF NOT EXISTS links (from_kind TEXT, from_id TEXT, to_kind TEXT, to_id TEXT,
   updated_at TEXT NOT NULL, PRIMARY KEY (from_kind, from_id, to_kind, to_id));
